pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Yenching Hsu",email = "Yenching_Hsu@asus.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain-community (>=0.3.30,<0.4.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "pypdf (>=6.1.1,<7.0.0)",
    "unstructured[pdf] (>=0.18.15,<0.19.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-openai (>=0.3.33,<0.4.0)",
    "chromadb (>=1.1.0,<2.0.0)",
    "langchain-chroma (>=0.2.6,<0.3.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "fastapi (>=0.118.2,<0.119.0)",
    "langchain-unstructured (>=0.1.6,<0.2.0)",
]

[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["rag_demo"]

